Add C++ solution for Point_Location_Test

+++ Geometry/Point_Location_Test/Point_Location_Test.cpp
@@ -12,7 +12,7 @@ int main(){
     int tt;
     cin >> tt;
     while(tt--){
-        int x1, y1, x2, y2, x3, y3;
+        long long x1, y1, x2, y2, x3, y3;
     cin >> x1 >> y1 >> x2 >> y2 >> x3 >> y3;
     if(((x2 - x1) * (y3 - y1)) == ((x3 - x1) * (y2 - y1))){
         cout << "TOUCH" << "\n";
